+++ tests/Fake/Ro/PathOnlyInputFixture.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BEAR\ApiDoc\Fake\Ro;
+
+use BEAR\Resource\ResourceObject;
+
+class PathOnlyInputFixture extends ResourceObject
+{
+    public function onPut(): ResourceObject
+    {
+        return $this;
+    }
+}